working popup!
This actually does away with the popup, and so provides the best of both
worlds: a popup that acts like a window, because it *is* a window.
The focus problem ended up just being that I should unclick "Non
Activating" in IB.

There's a slight flicker when you press the button, but I can live with
it for now. For better control, I basically need to implement the action
listener myself.