close menu popup when you click outside the window
As part of this, I made Atomic (formerly Atomic.Var) a full-blown atomic
reference holder, and not just a `@propertyWrapper`.

When you open the PTN window by pressing the menu button (but _not_ when
it pops up as a scheduled task), we use that Atomic class to register
a deactivation hook with AppDelegate. That hook closes the window on
deactivation.