Scheduled PTN popup (#27)
This reworks the popup menu to actually be a custom window (that looks a
lot like a menu!). That makes it easy to pop it up whenever we want.

The PTN popup starts on the visual foreground, but without focus. You
can focus it using ⇧⌘X.

This new fake menu approach also means we don't have to worry about the
stupidity of the activation sequence. (Activating the app closes the
menu bar, so before we had to handle activation as a two-step
process: we started with an empty menu that would then "open" but
show nothing, and then we would use that to activate, and then on
activation we would populate the menu and then pop it up. Now we
just show the window, and that's that.)

This resolves #5.

**Note**: This commit is amended from PR #27. That PR originally
referenced #6, not #5.