better code for version handling
We use `buildscripts/set_plist_build.sh` to get the abbreviated git
hash, which we then convert to decimal and put as the patch number. To
make this work, the public version has to always have the major and
minor versions, but not patch number. Then, if the git space is dirty,
we append `0xFFFF` to that hash (before we convert it to decimal).

In Version.swift, we reverse the process: we parse the version string's
components, and then treat the build number as a String (not decimal
number), which is the abbreviated hash, potentially with `"-dirty"` as a
suffix.