basics for developer release (#35)
This isn't a real release: we don't have the Apple groups thing yet,
Apple store stuff, etc. But I think it's enough so that you can
export a binary and download it somewhere else. We also target 10.14,
since that's the other computer I want to use this on. :-)

I also added some logging and version tracking.

This resolves #14. There's work to be done still for the real release,
but this is enough for now.