add helper to handle opening/closing the window
There are a few states we need to worry about:

- opening the window when it's closed (easy)
- opening a scheduled PTN while the PTN is already open
- ... while the daily report is already open
- opening the daily report while the daily report is already open
- ... while the PTN is open

Note that we don't need to worry about manually opening the window while
it's already opened; the UI won't allow that.

I think we can sum all of the above with a few easy rules:

- If window _A_ is open due to a schedule, or if someone tries to
  schedule another _A_ while _A_ is already open (for any reason), then
  schedule another _A_ when the window closes.
- If window _A_ tries to open when window _B_ is already open, then once
  _B_ is open, (1) schedule _B_ if needed by the above rule, and then
  immediately open _A_; and when _A_ later closes, schedule another _A_
  if needed by the above rule.

I think this will also generalize to other window types (like the daily
start), once we have those.

This is the generalization for the open logic that we wanted in #71.
I'll apply it to the actual UI in the next commit.