checkpoint: starting the autocomplete finding!
Unfortunately, I've just learned (via googling/SO) that it's impossible
to keep the popUp NSMenu open while also focusing on the text field; and
it's impossible to show the NSMenu another way. Instead, I have to
create my own menu bar-type view. I'm going to do that in the next
commit, so that I can more interactively debug and play around with the
rest of the matching logic.