fix autocompletes and text replacements
The primary motivator for this was that autocomplete wasn't triggering
on paste. For instance, if I had an empty field and I pasted `t`, it
would not autocomplete "two", since that pate wasn't a keyDown event.

As I dug into it, I realized that I had the pattern a bit wrong. I
shouldn't care what kind of event it is; I should just care what happens
to the selection. If there's a tail selection (including of length 0,
which just means the cursor is at the end), and the user extends into
it, then autocomplete. Otherwise, don't.

It's a bit complicated when you factor in other kinds of selection
events, like when a user manually selects all of the text and then
enters `t` (and wants `two` autocompleted). For that, you need to make
the field an NSTextViewDelegate and have it listen to
textViewDidChangeSelection. But that gets triggered before
textDidChange, so now you need to carry the "should I autocomplete"
state over to an instance variable.