fix the fix of focus :-)
The old way was breaking the daily report, because we would focus the
window (and thus lay it out) before we set the daily report as the
controller. I think that's literally the bug that I fixed last time I
changed this focus() call, but hey, that's what regression tests are
for 🤦.

To do this, I also had to pass the OpenReason down through to the
OpenCloseHelper.onOpen callback.