Fix a race condition in re-opening the window
If the PTN is open when the daily report is scheduled (or the other way
around), we want to re-open the window after it closes, with the new
conent. Prior to this change, that was racy: you could end up trying to
re-open the window before it had fully closed, which resulted in a noop.

Rather than try to beat this race (or sync it with hooks that don't seem
readily available), this commit avoids it. When we should re-open a
window, we instead just don't close it.

This also comes along with a lot of additional logging that I picked up
on the way, and figure we may as well keep.

This resolves #108.