don't close on deactivation
This wasn't happening consistently, anyways: the hook expires after the
first time it gets invoked, and the code wasn't refreshing it. None of
the tests expect it to happen. I did wonder a bit why it wasn't
happening, but I liked it so I didn't worry too much about it. Now I
understand. :-)