add DelegatingScheduler
The OpenCloseHelper will use this as a way to create a scheduler for
the PTN and daily report to use. They'll be able to schedule tasks like
"update such and such label in 5 minutes," and the OpenCloseHelper will
be able to easily cancel those tasks when it closes the window.