give OpenCloseHelper a DelegatingScheduler
OCH constructs a DelegatingScheduler on each call to its `opener`, which
it passes down to the open context (which is now a struct rather than
"loose" args, since I have more and more of those args). That scheduler
is only valid while the item is still open; it'll get closed when the
item's `didClose` gets called, even if a new window gets immediately
scheduled or re-opened.