make prefs generic and type-safe
This is in preparation of having more of them. I want to have a
nicely-typed, central repository of them.

It's a bit unfortunate that the readPref/writePref static methods leak
out of `Prefs.swift`, but oh well. Swift doesn't let me do a fully
fileprivate protocol-and-extension-to-conform.

Also, a minor tweaks to General prefs tab pane. Rather than having the
default y-align be "bottom", with the only non-empty row having a
"center" override, just make the default "center".