configurable PTN frequency
- Fixes #51: Creates a pref for the frequency and jitter, and hooks it
  up to the prefs box. We also reschedule the PTN on close if either the
  frequency or jitter changed.
- Fixes #148: I thought it would be easier to test the frequency and
  jitter work if closing the PTN also resets its timer. That didn't end
  up being the case per se, though the framework is the same (namely,
  the addition of the new `OpenCloseHelper#forceRescheduleOnClose`).

This commit also moves a chunk of the PtnViewControllerTest (the
"Check snooze option update" stuff). I forget exactly why I did that,
 but it made sense at the time.

Also also, make the mocked clock window a non-panel, but at the floating
layer. Hopefully that will help with test stability, as far as tests
that can't find that window.