build can notarize app
It won't for the automaetd builds, but it will if you request it on a
manual build. Notarization takes a while, so unless I end up needing it,
I don't want to do it each time. I could also do something where I
notarize on refs/heads/main, but don't wait for the result and staple it
unless requested (or maybe I always do? It only takes a few minutes in
practice). Basically, until I have a reason to notarize consistently,
it's not worth it to me to do it.

I was hoping this would fix the "launch on login" bug (namely, that it
works when I build in xcode but not via GH actions), but it doesn't
seem to.

Also some refactoring of the yml to combine "run" steps.