Fix window positioning with multiple screens
NSStatusBarButton's button.window doesn't seem to change until after it
displays, so I can't just rely on it to figure out where to open the PTN
or daily report window. Instead, I:

  1. Find the button bar's window
  2. Find how far that window is from its screen's right edge
  3. Find the screen that the mouse is currently in
  4. Calculate the point that's similarly far from that screen's right
      edge

This won't work for RTL screens, unfortunately. If a user has an RTL
configuration and one screen, then it doesn't matter much whether we
calculate the button's distance from the left or the right. But with
multiple screens, we want to calculate the distance from the left, since
that's where the user status items are anchored. I don't know how to get
at that information, so I filed #164 to follow up on it.

This fixes #90.