remove unneeded layout calls
We call these methods from within `animate()` anyway (which is always
what encloses the methods this diff changes), so we don't need to call
them here. Doing so incurs a big perf hit. With some random data I've
accumulated over the past couple months, setting the daily report's
start to June took ~1.66 sec before, and now takes about 0.467 sec.

This obviates a lot of the need for #146, since the daily report is now
about 4x faster (and we therefore don't benefit as much from a progress
bar while it's going).