consolidate test controls (#205)
* consolidate test controls

Rather than two windows (one for components, and one for the mocked
clock), just have the one window and move the mocked clock into it.

This lays the groundwork for #189 (an automated screenshot creator).
As part of that, I'm going to want to move the PTN's entries json into
what was the mocked clock, so that the PTN itself always looks normal,
even during UI testing. But that means that the mocked clock is
increasingly doing more than just being a mocked clock (that, reset,
soon the entries), so this gets us towards that goal.

As part of that, move entries-json hook to main mocked UI helper.

It was in the PTN, but we're moving it out. This will unblock #189 by
letting me take screenshots of the PTN without that hook field getting
in the way.