sizing fixes for day-end report (#211)
This fixes #210, though not in a super great way. I fixed the
"permanent" issues (views growing too large, and then not shrinking
again), but the transitions are pretty iffy. But, I think it's good
enough for 1.0.

Included in this is a refining of the UI hook's animation knob. Rather
than just being a boolean, it's a double now, which is the speed factor
to apply to all animations. Each animation will take `k * d` seconds,
where `d` is its stated duration in the code, and `k` is the factor; so
to effectively disable animations, set `k = 0`.

Also:
- remove the version string from the day-end report; we don't need it
  there, now that it's in the prefs pane
- test improvement: `Existence` tri-value enum to just boolean isVisible
- test fix: correctly restore the status item position after a test
  failure