day-end styling
This moves the horizontal divider line that was 4 pixels above the
scroll to be immediately above it, and adds one below it. This
effectively is the same as removing ithe lines and making the scroll
view bordered, but I couldn't figure out how to do that well: when I
tried, it made the view too small by a few pixels, and thus kept the
scroll bar always there.