improved logging
This is to help understand our flakey tests (#209). With this change, we
log more correctly (using `os_log` instead of `NSLog`, the latter of
which is actually meant just for warnings), but also fork those log
messages over to the app control window in UI tests. This lets us copy
them and add them as an attachment, which should make it easier to debug
failures.