fix tests
Before, we were using endSheet(NSWindow) to close the sheets when
resetting, which sends a .stop signal. If the prefs dialogue is open,
that results in an app termination. That happens in the setUp for tests,
causing them to fail due to a closed connection. So, call the overload
to explicitly set a .cancel instead, which won't cause us to quit the
app.

Also, use `click(using: .frame())` instead of `click()` in one spot. The
normal click works fine on GH Actions, but fails for me locally because
the app isn't active.