daily report is always one day at a time
Previously, if you selected a past day, it would give you all of the
stuff from that previous day to now. With this commit, it will just give
you that previous day.

This makes it easier to quickly see "what did I do yesterday?", which
is what I usually want from this.

For more long-term reports, I'll probably want an option to explicitly
specify a range.