improve date range picker in day-end report (#237)
Rather than selecting a start date and a assuming an end date of
"tomorrow morning", let the user select both. This comes with a more
sophisticated date-range picker that is possibly nicer to use, and
definitely easier to UI-test.

Also various improvements to the UI tests. They're failing on GH, but
working locally; I think it's because GH has downgraded the machines, at
least for free-tier usage.