window for seeing larger task summaries
This is similar to the day-end report, but it can handle more items.
It can also sort them in various ways. The hope is that this is useful
for "what did I do this past year" style reporting, and also for finding
infrequently-used projects and tasks that could be consolidated into
others.

There's more work to be done here (including automated testing), but
this is a good milestone!

This is part of the way towards #269 (though it still needs testing
before I can call it done).