only look for "Build app" jobs on main
If I have a feature branch `foo` with a PR and then ff-merge that to
main, then there'll be two "Build app" jobs for that sha: one from the
PR, and one from main. I only want one of them, or else the
`build_run_id` var will have both, separated by a newline (which is not
a valid build run id)! And it should specifically be the one from
`main`, since that's the one that will have gotten notarized.