search functionality across all projects and tasks (#280)
Press `⌘F` to open up a "find" field in the PTN
window. This lets you search across all projects
and tasks, using the same autocomplete as the
project and task fields. The segmentation of
projects and tasks within the field is slightly
hacky, but it's good enough. :-)

I need to test this in the field to get a sense of
whether it's performant enough. In the meanwhile,
I'm going to keep the tracking ticket (#114) open.
If this ends up being acceptable, I'll close the
ticket.