refactoring
Pull all the string building out to a class; this will let us recreate
it for expected values in testing.

Also, minor: change the second segment in the UiTestWindow component
from `two` to `2`. This is shorter than `one`, and will thus have
precedence when autocompleting. This will be useful for tests.