xcode build setting updates; gh action updates
I don't know what the xcode project changes are, but xcode recommended
them so I'm going with that. :-)

But as part of it, it moved around which certs it wants me to build
with. While I was there, I also changed the notarization tool from
altool to notarytool. This fixes #305.